Add an array struct interface and the ndim, itemsize, shape, strides, and _pixels_address properties.
Fix problems with slicing when the step was not 1. It would incorrectly calculate the new array's size in some cases.

Allow for 2D PixelArray's with at least one dimension having length 1. This was not possible before, as a length 1 dimension was automatically removed, reducing the array's dimension by 1. 1D arrays are still possible, when an integer index is used. This is consistent with NumPy.

Rework the PixelArray internals to use the data pointer, shape, and strides array model used by NumPy and exported by the array struct and new buffer interfaces. The code is more compact, limits some of the tricky calculations to only a few places, and more closely resembles Surface blitting.

--HG--
extra : convert_revision : svn%3A6368f583-b015-0410-a43b-8a248398de60/trunk%403126