Remove experimental _freetype.Face methods
Methods get_transformation, set_transformation, and delete_transformation
were part of an attempt to generalize the rotation and oblique transforms.
The provided an interface to a common transformation matrix. As it is,
the rotation and oblique transforms involve more than than just applying
a matrix multiplication on a glyph outline. This is too complicated
to be worth exploring right now. It would better be developed in a
Pygame branch.