Automatically generate the version module with setup.py
The version.py module is now written by setup.py during a build. This means,
for the Pygame code, only the 'version' tag within setup.py needs to be
changed. Also, a revision number, 'rev', is added to the version module.
This gives the hg identifier of the build. It is automatically determined
by setup.py.

--HG--
rename : lib/version.py => version.py.in