improve pygame.freetype.Font font size handling (towards Issue #75)
These changes may cause minor breakage, as the *ptsize* argument to the
pygame.freetype.Font constructor and several of its methods is renamed *size*.

Perform better error checking on the *size* (formerly *ptsize*) argument, and
use this size, now *face_size* within the C code, more consistently. Its
type now matches the width and height fields of the FTC_ScalerRect C struct.
The face size is a value expressed in 1/64th or points. It is a transition
to allowing a two-tuple of floats as Python *size* value, which separates the
width and height requirements.